requirements.txt
docx2python==2.8.0
fastapi==0.109.0
ipython==8.13.2
LexRank==0.1.0
nltk==3.8.1
numpy==1.24.3
pandas==2.0.1
pyfiglet==0.8.post1
python-dotenv==1.0.1
python_Levenshtein==0.25.0
quart==0.19.5
quart_cors==0.7.0
Requests==2.31.0
sentence_transformers==2.2.2
skillNer==1.0.3
skillNer==1.0.3
spacy==3.5.3
toml==0.10.2
torch==2.1.2
txtai==5.5.1
uvicorn==0.27.0.post1
requests==2.31.0
